requirements.txt
praw==7.7.1
python-dotenv==1.0.1
torch==2.3.0+cu118
transformers==4.40.0